elasticsearch integration test
