+++ src/test/java/com/solvd/micro9/users/integration/elastic/ElasticTestcontainers.java
@@ -8,16 +8,19 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
 @Testcontainers
-abstract public class ElasticTestcontainers {
+public abstract class ElasticTestcontainers { //TODO remove PUBLIC
 
     private static final String TEST_PASSWORD = "testPassword";
 
     @Container
-    public static ElasticsearchContainer container = new ElasticsearchContainer(
-            DockerImageName
-                    .parse("elasticsearch:8.7.0")
-                    .asCompatibleSubstituteFor("docker.elastic.co/elasticsearch/elasticsearch")
-    )
+    private static final ElasticsearchContainer ELASTICSEARCH_CONTAINER =
+            new ElasticsearchContainer(
+                    DockerImageName
+                        .parse("elasticsearch:8.7.0")
+                        .asCompatibleSubstituteFor(
+                                "docker.elastic.co/elasticsearch/elasticsearch"
+                        )
+            )
             .withPassword(TEST_PASSWORD)
             .withEnv("discovery.type", "single-node")
             .withEnv("xpack.security.enabled", "true");
@@ -25,7 +28,7 @@ abstract public class ElasticTestcontainers {
     @DynamicPropertySource
     static void kafkaProperties(final DynamicPropertyRegistry registry) {
         registry.add("spring.elasticsearch.url",
-                container::getHttpHostAddress
+                ELASTICSEARCH_CONTAINER::getHttpHostAddress
         );
         registry.add("spring.elasticsearch.password", () -> TEST_PASSWORD);
     }
